Fix#1142 datasource lookup with tomcat jndi prefix
